move dns_lru to separate module
