+++ resolver/src/lib.rs
@@ -138,6 +138,7 @@ extern crate trust_dns_proto;
 extern crate ipconfig;
 
 pub mod config;
+mod dns_lru;
 pub mod error;
 pub mod lookup_ip;
 pub mod lookup;
